requirements.txt
anyio==3.6.2
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
dnspython==2.3.0
exceptiongroup==1.1.1
fastapi==0.95.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
mypy==1.1.1
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pluggy==1.0.0
pydantic==1.10.6
Pygments==2.14.0
pymongo==4.3.3
pytest==7.2.2
python-dotenv==1.0.0
PyYAML==6.0
rfc3986==1.5.0
rich==12.6.0
shellingham==1.5.0.post1
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typer==0.7.0
typing_extensions==4.5.0
uvicorn==0.21.1
watchfiles==0.18.1
websockets==10.4
